requirements.txt
python-dotenv~=1.0.0
pydub~=0.25.1
tiktoken==0.3.3
openai
python-telegram-bot
requests~=2.30.0